# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Canvas API",
    "version": "1.0.0",
    "description": "Локальный API для рабочего пространства с нодами. Авторизация не нужна. GET графа возвращает ETag; PUT требует If-Match. Ответы содержат сам ресурс, без обёртки data. Генерация имитируется."
  },
  "components": {
    "schemas": {
      "def-0": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "additionalProperties": false,
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        },
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/api/config": {
      "get": {
        "operationId": "getConfig",
        "summary": "Параметры тестового",
        "tags": [
          "API"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "if-match",
            "required": false,
            "description": "Условие для ресурса по адресу запроса. ETag выдаётся только там, где указан в ответе."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "description": "Совпадение: 304 для GET/HEAD, 412 для изменения."
          }
        ],
        "responses": {
          "200": {
            "description": "Текущее представление ресурса.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "debounceMs",
                    "pollIntervalMs",
                    "generationDelayMs",
                    "maxNodes",
                    "maxEdges",
                    "nodeTypes",
                    "links"
                  ],
                  "properties": {
                    "debounceMs": {
                      "type": "integer"
                    },
                    "pollIntervalMs": {
                      "type": "integer"
                    },
                    "generationDelayMs": {
                      "type": "integer"
                    },
                    "maxNodes": {
                      "type": "integer"
                    },
                    "maxEdges": {
                      "type": "integer"
                    },
                    "nodeTypes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "links": {
                      "type": "object",
                      "additionalProperties": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "href",
                          "method"
                        ],
                        "properties": {
                          "href": {
                            "pattern": "^/",
                            "type": "string"
                          },
                          "method": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "GET"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "POST"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "PUT"
                                ]
                              }
                            ]
                          }
                        }
                      }
                    }
                  },
                  "description": "Текущее представление ресурса."
                }
              }
            }
          },
          "304": {
            "description": "Условие If-None-Match совпало; без тела.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные поля, JSON или заголовки.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "412": {
            "description": "PRECONDITION_FAILED: условие запроса не выполнено.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      },
      "head": {
        "operationId": "head_getConfig",
        "summary": "Заголовки: Параметры тестового",
        "tags": [
          "API"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "if-match",
            "required": false,
            "description": "Условие для ресурса по адресу запроса. ETag выдаётся только там, где указан в ответе."
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "description": "Совпадение: 304 для GET/HEAD, 412 для изменения."
          }
        ],
        "responses": {
          "200": {
            "description": "Текущее представление ресурса.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Условие If-None-Match совпало; без тела.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные поля, JSON или заголовки.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "412": {
            "description": "PRECONDITION_FAILED: условие запроса не выполнено.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR.",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "options": {
        "operationId": "options_getConfig",
        "tags": [
          "HTTP"
        ],
        "summary": "Разрешённые методы",
        "responses": {
          "204": {
            "description": "Без тела.",
            "headers": {
              "Allow": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}